move LazyHDF5 class into its own file to resolve circular imports
